refactor(frontend): extract FilterDrawer component from titles page
Move the mobile filter drawer (toggle button, backdrop, slide-in panel,
scroll lock, escape key, focus management) into a reusable component.
